+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.7] - 2023-01-06
+
+### Changed
+
+- Refactored all unit tests
+
+### Fixed
+
+- All code is now C# 7.3 and .NET Framework 4.6 compatible for Unity 2019.1+
+
 ## [0.4.6] - 2023-01-04
 
 ### Added